+++ frontend/src/components/Chat.js
@@ -5,14 +5,7 @@ export default function Chat(props) {
     const { query, answer } = props;
     const [chats, setChats] = useState([answer]);
 
- 
- 
-    function keyPress(e){
-        if(e.key === 'Enter'){
-            e.preventDefault();
-            setChats([...chats, e.target.value]);
-        }
-    }
+
 
     return(
         <>
@@ -31,24 +24,7 @@ export default function Chat(props) {
             ))}
         </Stack>
         </Paper>
-    
 
-        <Paper
-            component="form"
-            position="fixed"
-            sx={{display: 'flex', position: 'fixed',
-            bottom: 50,
-            width: 900, 
-            alignItems: 'center', 
-            height: 70, border:1, borderColor: "#DFE1E5"}}
-        >
-        <InputBase
-            sx={{ ml: 1, flex: 1, fontSize: 20, m: 3}}
-            onKeyDown={keyPress}
-            placeholder="Send A Message"
-            inputProps={{ 'aria-label': 'search google maps' }}
-        />
-        </Paper>
         </>
     )
 }
