+++ frontend/src/pages/SearchResult.js
@@ -1,5 +1,5 @@
 import { useSearchParams } from 'react-router-dom'
-import { Typography, Box, Grid } from '@mui/material';
+import { Typography, Box, Grid, Paper, InputBase, Stack } from '@mui/material';
 import SourceList from '../components/SourceList';
 import Chat from '../components/Chat';
 import { useEffect, useState } from 'react';
@@ -30,24 +30,65 @@ export default function SearchResult() {
           };
     }, [searchParams])
 
+    function keyPress(e){
+        if(e.key === 'Enter'){
+            e.preventDefault();
+            // setChats([...chats, e.target.value]);
+        }
+    }
+
 
     return(
         <>
         <Box sx={{ flexGrow: 1 }}>
-        <Grid container spacing={2}>
-            <Grid item xs={2}/>
-            <Grid item xs={6} sx={{mt:5}}>
-                { q &&
-                <Chat query={q} answer={responseMessages.map(mes => mes.chat_response).join('')}/>}
+            <Grid container spacing={2}>
+                <Grid item xs={2}/>
+                <Grid item xs={6} sx={{mt:5}}>
+                     {/* Chat Messages */}
+                    <Paper sx={{height: 700, overflow: 'auto'}} elevation={0}>
+                        <Stack sx={{ mr: 6}} height='50px'>
+                            <Typography variant="h3" gutterBottom>
+                                {q}
+                            </Typography>
+                            
+                            <Box sx={{borderBottom:1, pb: 3, mb: 3}}>
+                                <Typography variant="body1" fontSize='20px' mr={6}>
+                                 {responseMessages}
+                                </Typography>   
+                            </Box>    
+                           
+                        </Stack>
+                    </Paper>
 
-            </Grid>
-            {responseMessages.length > 0 &&
-                <Grid item xs={4}>
-                    <SourceList source={responseMessages} urls={responseMessages[0].documents.map(doc => doc.metadata.url)}/> 
+                    {/* Chat Bar */}
+                    <Paper
+                        component="form"
+                        position="fixed"
+                        sx={{display: 'flex', position: 'fixed',
+                        bottom: 50,
+                        width: 900, 
+                        alignItems: 'center', 
+                        height: 70, border:1, borderColor: "#DFE1E5"}}
+                    >
+                        <InputBase
+                            sx={{ ml: 1, flex: 1, fontSize: 20, m: 3}}
+                            onKeyDown={keyPress}
+                            placeholder="Send A Message"
+                            inputProps={{ 'aria-label': 'search google maps' }}
+                        />
+                    </Paper>
+
                 </Grid>
-            }
-        </Grid>
+                {responseMessages.length > 0 &&
+                    <Grid item xs={4}>
+                        <SourceList source={responseMessages} urls={responseMessages[0].documents.map(doc => doc.metadata.url)}/> 
+                    </Grid>
+                }
+            </Grid>
         </Box>
+
+
+       
         </>
     )
 }
